+++ src/figcli/config/communication.py
@@ -82,6 +82,8 @@ PROFILE_HELP_TEXT = "Overrides all other figgy configurations and instead pulls 
 IAM_RESTORE_HELP_TEXT = "Restores any credentials in ~/.aws/credentials that were automatically backed up by figgy."
 UI_HELP_TEXT = "Loads the Figgy UI in your default browser."
 RUN_HELP_TEXT = "Runs the Figgy UI -- this command is optional and can be run by default with `figgy ui`. "
+BUILD_CACHE_HELP_TEXT = "Makes figgy log-in to every account/region and build out a local cache for uber fast figgy usage. The cache will auto-build on its own with usage."
+
 
 # Point in time (--point-in-time)
 POINT_IN_TIME = "Restore all parameters to a point in time."
@@ -132,7 +134,8 @@ HELP_TEXT_MAP = {
     upgrade: UPGRADE_HELP_TEXT,
     iam_restore: IAM_RESTORE_HELP_TEXT,
     ui: UI_HELP_TEXT,
-    run: RUN_HELP_TEXT
+    run: RUN_HELP_TEXT,
+    build_cache: BUILD_CACHE_HELP_TEXT
 }
 
 # Other